Refactor folder structure (#9)
